separate Item and CustomAdapter
