+++ app/src/main/kotlin/jp/co/yumemi/android/codeCheck/domain/Item.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.codeCheck
+package jp.co.yumemi.android.codeCheck.domain
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
