+++ app/src/main/kotlin/jp/co/yumemi/android/codeCheck/ui/CustomAdapter.kt
@@ -1,0 +1,36 @@
+package jp.co.yumemi.android.codeCheck.ui
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
+import jp.co.yumemi.android.codeCheck.domain.Item
+import jp.co.yumemi.android.codeCheck.R
+import jp.co.yumemi.android.codeCheck.diff_util
+
+class CustomAdapter(
+    private val itemClickAction: ItemClickAction,
+) : ListAdapter<Item, CustomAdapter.ViewHolder>(diff_util) {
+
+    class ViewHolder(view: View) : RecyclerView.ViewHolder(view)
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
+        val view = LayoutInflater.from(parent.context)
+            .inflate(R.layout.layout_item, parent, false)
+        return ViewHolder(view)
+    }
+
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
+        val item = getItem(position)
+        (holder.itemView.findViewById<View>(R.id.repositoryNameView) as TextView).text =
+            item.name
+
+        holder.itemView.setOnClickListener {
+            itemClickAction(item)
+        }
+    }
+}
+
+typealias ItemClickAction = (item: Item) -> Unit
